requirements.txt
streamlit>=1.35.0
anthropic>=0.18.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
watchdog>=3.0.0
pillow>=10.2.0 